generate flat dialogue file from tree
