+++ game/dialoguesApi.js
@@ -1,3 +1,3 @@
-import { dialogues } from "./dialoguesFlat";
+import dialogues from "./dialoguesFlat.json";
 
 export const getNode = (nodeId) => dialogues[nodeId];
